Fixed conflicting issues with main activity and updated master branch
