+++ app/src/main/java/com/example/connectsit/ui/model/Enterers.kt
@@ -1,4 +1,4 @@
-package com.example.connectsit.model
+package com.example.connectsit.ui.model
 
 enum class Enterers(val value: String) {
     STUDENT("STUDENT"),
