--- app/src/main/java/com/example/connectsit/ui/screen/ScreenUtils.kt
@@ -1,15 +1,0 @@
-package com.example.connectsit.ui.screen
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-object TeacherScreenRoute
-
-@Serializable
-data class LogInScreenRoute(val name:String)
-
-@Serializable
-object OptionChoosingScreenRoute
-
-@Serializable
-object StudentPortalRoute
