--- app/src/main/java/com/example/connectsit/ui/screen/student/StudentPortal.kt
@@ -1,12 +1,0 @@
-package com.example.connectsit.ui.screen.student
-
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-
-@Composable
-fun StudentPortal(){
-
-    Text(text = "Student portal")
-
-}
