+++ app/src/main/java/com/example/connectsit/ui/screens/auth/StudentTeacherDeterminerScreen.kt
@@ -1,4 +1,4 @@
-package com.example.connectsit.ui.screen.auth
+package com.example.connectsit.ui.screens.auth
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
@@ -17,12 +17,15 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
-import androidx.navigation.NavHostController
 import com.example.connectsit.R
-import com.example.connectsit.ui.screen.LogInScreenRoute
+import com.example.connectsit.ui.model.Enterers
 
+// moved this composable to its own file to remove clutter in MainActivity
 @Composable
-fun OptionChoosingScreen(navController: NavHostController) {
+fun StudentTeacherDeterminerScreen(
+    // create anonymous function to pass back navigation to MainActivity
+    handleNavigation: (Enterers) -> Unit
+) {
     Column(
         modifier = Modifier
             .fillMaxSize()
@@ -40,11 +43,8 @@ fun OptionChoosingScreen(navController: NavHostController) {
         Spacer(modifier = Modifier.height(16.dp))
         Button(
             onClick = {
-                navController.navigate(
-                    LogInScreenRoute(
-                        name = "STUDENT",
-                    )
-                )
+                // Created Enum to safely determine which user/enterer we are displaying for
+                handleNavigation(Enterers.STUDENT)
             },
             colors = ButtonDefaults.buttonColors(
                 containerColor = Color.Blue,
@@ -58,11 +58,8 @@ fun OptionChoosingScreen(navController: NavHostController) {
 
         Button(
             onClick = {
-                navController.navigate(
-                    LogInScreenRoute(
-                        name = "TEACHER",
-                        )
-                )
+                // Created Enum to safely determine which user/enterer we are displaying for
+                handleNavigation(Enterers.TEACHER)
             },
             colors = ButtonDefaults.buttonColors(
                 containerColor = Color.Blue,
